Fundamental pagination flaw fixed with re-implementation, with other fixes.
After attempting to do some heavy local caching of a GitHub org (local
processing is -much- faster than trying to do remote requests, let alone having
to be concerned with quotas and whatnot), there would be times where the tool
would just -hang-.  No new output was produced, but no network or disk activity
was active.  And of course this appeared to be intermittant...

THE SYMPTOMS AND TROUBLESHOOTING:

The main way that this failure showed itself was by periodically stalling.  A
quick run through with 'set -x' set showed that it was failing at the final
step, where multiple json files are stiched together.  For some reason, the
list of files to stitch was coming up empty.  Feed no input to a tool that is
expecting input, and you get a hanging tool.

Figuring out -why- the files were dissapearing was much harder than expected.
At first I though that the files weren't being created, or that requests
weren't being made.  At one point, I confirmed that files were dissapearing
before they were attempting to be merged, which did help isolate the problem to
the substituted process, but for the longest time it seemed like it was humming
along just fine.

At one point, I attempted to simplify the everything by removing the final 'sed'
command within the curl request, and was reminded why this implementation ended
up like this in the first place.  Without the final 'sed', the rest of the
script continues as if nothing happened and eventually completes.  Meanwhile,
the process substitution almost always takes much longer, and when it finally
does finish, the files are immediately cleaned up due to the trap, and you're
left with only the first page of data.

While debugging this issue, it became trivially apparent when the terminal
prompt would reappear, in the middle of the substituted process still doing its
thing.

THE ROOT CAUSE:

As I -now- know, traps are -not- inherited by subshells, but -are- with process
substitution.  And whenever a --paginated request was made, when it turned out
that no additional pages would be needed, the 'seq' would fail since 'seq 2 1'
is understandably not valid.

This then triggers the 'cleanup' trap, resulting in that entire string of
commands capturing the first request, then blowing it away because the final
'sed' command has long since stopped getting input.

As for why this appeared intermittant, welp, -that's- because before I realized
that -this- was the problem, I was trying other things like varying the
'per_page' size, and the request URL, -both- of which changed the behavior of
the problem...

THE RE-IMPLEMENTATION:

The nice thing about all of this was that in my time away from managing this
block of code, I gathered some new tools.  I remembered that 'curl' supports
outputing only the header into a file, and used this with a temporal fifo to do
the header analysis (in fact, I think I originally tried to do this -without-
a fifo, for reasons only a younger me only knows).  Once this was done, I could
get rid of the weird implementation in its entirety since I can throw the first
'curl' request into the background, read the fifo to determine how many more
requests to make, make them in the background, and wait on all of the requests
to resolve within the main script.

Oh, and I made sure that 'seq' doesn't try to run if only the first page is
needed.

Now, for the minior fixes:
- Always request the max datum size, even for simple requests
- Disabled datum check for 'get' request, since it can apparently accept data...
- Added trap for user interruption.  Useful for those really large requests.