Fixed how URL parameters are handled.
Found out that some requests require additional URL parameters to be set, not payload to be added.
Before, there was no way to append extra URL params since '?per_page=100' was always being set.

Now, the URL is split and tested to see if the user is passing any additiona params.
The URL is then rebuilt with the additional 'per_page=100' param tacked on.